pyproject.toml
[project]
name = "ingestion-workflow"
version = "0.1.0"
description = "Modular ingestion workflow for Neurostore data pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Ingestion Workflow Team" }
]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.5",
  "requests>=2.31",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "pandas>=2.2",
  "lxml>=4.9",
  "pubget @ git+https://github.com/neuroquery/pubget.git@main",
  "pydantic-settings>=2.5",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2",
  "pytest-cov>=4.1",
  "pytest-recording>=0.13",
]

[project.scripts]
ingest = "ingestion_workflow.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ingestion_workflow*"]
